Allow setting floating point numbers as range and step size

+++ app/src/main/java/it/sephiroth/android/numberpicker/demo/MainActivity.kt
@@ -43,7 +43,7 @@ class MainActivity : AppCompatActivity() {
     }
 
     inner class Presenter {
-        fun onProgressChanged(numberPicker: NumberPicker, progress: Int, fromUser: Boolean) {
+        fun onProgressChanged(numberPicker: NumberPicker, progress: Float, fromUser: Boolean) {
             Timber.d("onProgressChanged")
             model.globalProgress.value = progress
         }
